requirements.txt
fire==0.6.0
torch==2.2.0
torchvision==0.17.0
pillow==10.2.0
matplotlib==3.6.3
accelerate==1.3.0
bitsandbytes==0.45.1
transformers==4.48.1